frontend not compiling with frontend-common (#17)
