+++ .gitignore
@@ -10,6 +10,8 @@ coverage
 
 # production
 /build
+build
+lib
 
 # misc
 .DS_Store
