+++ frontend-plugin/package.json
