+++ frontend-plugin/public/manifest.json
