+++ frontend-plugin/src/example.component.tsx
