+++ frontend-plugin/src/react-app-env.d.ts
