+++ lerna.json
@@ -1,6 +1,7 @@
 {
   "packages": [
-    "packages/*"
+    "packages/*",
+    "frontend-plugin"
   ],
   "version": "0.0.0",
   "npmClient": "yarn",
