+++ package.json
@@ -4,6 +4,7 @@
   "version": "1.0.0",
   "workspaces": [
     "backend",
+    "frontend-plugin",
     "packages/*"
   ],
   "description": "This is the COMP3000 Computing Project on the topic of creating \"A Web Platform for Visualising Optimisation Data\".",
